requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
qdrant-client==2.7.0
openai==1.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
tqdm==4.66.1
pydantic==2.5.0
requests==2.31.0
